refactor frontend app structure
